+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * Remove Events requirements, beem.notify and beemapi.websocket, as it is not well tested and there are no websocket api nodes available on hive
 * Remove unnecessary requirements (pylibscrypt and future
 * add new node (fin.hive.3speak.co) and change change rpc.esteem.app to rpc.ecency.com
+* Replace diff_match_patch by difflib and add unit tests
 
 0.24.13
 -------